add LutronControllerBaseEntity
